renamed messageSerializer to serializer, renamed made serialize a generic method, can take any object
added de-serialize account method

added fix to clientConnection, not a perfect fix but mostly fixes the issue

I'm on my laptop, and I'm not completely set up, but I wanted to get this done. If someone could pull this down and just do a quick check that would be great.